tests/wpt/web-platform-tests/css/css-fonts/parsing/font-synthesis-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-synthesis with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-synthesis-prop">
<meta name="assert" content="font-synthesis supports the full grammar 'none | [ weight || style ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-synthesis', 'none');
test_valid_value('font-synthesis', 'weight');
test_valid_value('font-synthesis', 'style');
test_valid_value('font-synthesis', 'style weight', 'weight style');
</script>
</body>
</html>
